reduce code length on strava page and separate out into other pages - organize directory in page folders
