+++ lib/Pages/competition_lobby.dart
@@ -5,9 +5,9 @@ import 'package:google_fonts/google_fonts.dart';
 import 'package:ride_tide_stride/components/competition_dialog.dart';
 import 'package:ride_tide_stride/components/competition_learn_more.dart';
 import 'package:ride_tide_stride/components/passwordDialog.dart';
-import 'package:ride_tide_stride/pages/mtn_scramble_page.dart';
-import 'package:ride_tide_stride/pages/snow_2_surf_page.dart';
-import 'package:ride_tide_stride/pages/team_traverse_page.dart';
+import 'package:ride_tide_stride/screens/challenges/mtn_scramble/mtn_scramble_page.dart';
+import 'package:ride_tide_stride/screens/challenges/snow2surf/snow_2_surf_page.dart';
+import 'package:ride_tide_stride/screens/challenges/team_traverse_page.dart';
 
 class CompetitionLobbyPage extends StatefulWidget {
   const CompetitionLobbyPage({super.key});
